Do not check for OpenCL device
The configure logic tests just the first icl for a valid device. If it doesn't have a CPU device (e.g. NVIDIA icd), the test fails and opencl tests are disabled. This patch removes the check and relies on pencil-runtime to select a working device. Since the ppcg command line option to select a device have been removed, it wasn't working anymore.

Signed-off-by: Michael Kruse <ppcg@meinersbur.de>