Hide pencil-util from automake
Automake use the conditional SUBDIRS to construct DIST_SUBDIRS. The are recoursed into during make check, which may fail if they have not been configured because they are not in SUBDIRS.  This patch ensures that pencil-util is not added to
the distribution when configured with something else than --with-pencil-util=bundled.  pencil-driver ships pencil-util on its own, so a second pencil-util would be redundant.