gpu.c: extract_single_tagged_access: clean up space copying
The copy of space is needed by isl_space_domain,
but it was being assigned to space2, which is subsequently overwritten.
In practice, this happens to work out because taking
a copy only changes the reference count and returns
the same pointer, but the original code is still confusing,
if not technically incorrect.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>