try and shift bands to the origin before tiling
If the tiled band does not start at the origin
then the initial tiles are likely to be partial tiles.
Shifting to the origin improves the changes
of starting with full tiles.
Do not perform any shift if this would involve piecewise expressions
since doing so is likely to result in more complicated code.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>