gpu.c: insert_empty_permutable_band: set domain of partial schedule
This ensures that the shifting introduced in the previous commit
can also be applied to these zero-dimensional bands.
In particular, shift_to_origin calls isl_schedule_node_band_shift,
which checks that the domain of the partial schedule
of the band node is not affected by the shift.
This assumes, however, that the partial schedule
has a domain to begin with.  Since insert_empty_permutable_band
did not originally set the domain, calling shift_to_origin
on such a node would result in a failure.

An alternative approach would be to special case
zero-dimensional band nodes in shift_to_origin,
or even to skip the shifting if the shift is obviously zero.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>