gpu: avoid mapping some nested non-permutable bands to the device
If the outer node of the schedule tree is a set or a sequence,
then all or the initial and final children that do not have
any permutable bands are already executed on the CPU.
The outer node may however have children that are themselves
set or sequence node.  If only some of their children have
permutable bands, then the entire child of the outer node
was still being mapped to the device.
Allow initial/final children to be collected recursively
to avoid more non-permutable bands getting mapped to the device.
In the case of a set node, recursion can be performed on any child
since the initial/final parts of any child of a set node can
be moved first/last.
In the case of a sequence node, recursion can only
be performed on the first/last child.

Note that the entire set of collected descendants is moved
before/after the other descendants.  The relative order
within each of these two parts is preserved.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>