gpu.c: add_to_from_device: pass down domain to node_may_persist
The domain has already been computed by the caller of add_to_from_device
so it does not need to be recomputed inside node_may_persist.

Note that there is a call to mark_kernels in between,
which may group statements and cause the expanded domain
in node_may_persist to not have all original domain constraints.
That is, the domain passed down may be a subset
of the originally computed domain, but the extra elements
are irrelevant because they do not belong to the original
statement instance sets.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>