take into account live-out writes while computing may-persist set
For determining which elements need to be copied into the device
the set of elements that may need to be preserved by
the selected subtree is computed since the copy-out operation
will overwrite these elements.
This may-persist set consists of the elements
that may need to be preserved by the entire analyzed fragment and
those that are in potential dataflow across the selected subtree.

There is, however, a third class of elements that may need
to be preserved and those are the ones are effectively written
by the analyzed fragment, that are need after this fragment and
that may get overwritten by a copy-out operation.

For example, in the new test cases, an element of the b-array
is written in a statement that does not get mapped
to the device and that is executed before the statement
that is mapped to the device.  Since this statement
also writes to (different elements of) the b-array,
this array get copied-out and overwrite the original element
if the b-array is not copied in first.

Add those elements to the may-persist set as well
so that they will be taken into account while computing
the copy-in set.

Note that the outer_may_overwrite field of ppcg_may_persist_data
could also be replaced by its domain.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>