gpu.c: get_non_parallel_subtree_filters: select subtrees based on node type
The function get_non_parallel_subtree_filters selects either
the initial non-parallel subtrees or all non-parallel subtrees.
Originally, the choice was made based on a function parameter,
but in the calls of this function this parameter is uniquely
determined by the node type.
Directly use the node type inside get_non_parallel_subtree_filters.
This makes it easier to introduce a function parameter
to choose between the initial and the final subtrees
(for sequence nodes) in the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>