gpu.c: add_to_from_device: store copy-out relation in copy_out
The variable may_write is reused for different kinds of relations,
one of which is constructed starting from the copy-out relation,
which is itself constructed from another may-write relation.
The copy-out relation was therefore also stored in may_write and
then copied into copy_out.  Store the copy_out relation directly
in copy_out and then copy that to may_write instead.
This clarifies what the object now stored in copy_out represents
at the point of the call to node_may_persist,
to which it will be passed in an upcoming commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>